Create cardmagic
This game allows up to 5 players. The players will be provided with a pile of 5 random cards each and are to think of a card from that randomized pile. cardmagic will allow the players to interact with the cards virtually and shuffle these cards, and at the end of the game, cardmagic will accurately predict the card that each player thought of.